Fix CLLocationManager stop & start stubbing
At the moment when `CLLocationManager` is stopped, the delegate it
removed from `_instanceHashTable` and later when manager is started
again, Swizzles tries to find the delegate from the same table but it is
not there.
To fix that issue, store all delegates on separate `_delegatesHashTable`
and remove only from `_instanceHashTable`. This way it is possible to
access the delegate later when manager resumes.